% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_tract.R
\name{census}
\alias{census}
\title{Census Tract}
\usage{
census(read_database, sql_query, write_database)
}
\arguments{
\item{read_database}{A database name}

\item{sql_query}{A SQL query (ex: select * from ...)}

\item{write_database}{A database name}
}
\description{
read_database : The database that the data is expected to read. \cr \cr
sql_query : A SQL query to read the table. \cr \cr
write_database : The database that the output is expected to be loaded. \cr \cr
}
\examples{
census(
 read_database = xyz,
 sql_query = "select * from table name",
 write_database = xyz,
)
}
