% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbd_batch_geocoding.R
\name{tbd_batch_geocoding}
\alias{tbd_batch_geocoding}
\title{Internal Function for tbd_address_to_tract}
\usage{
tbd_batch_geocoding(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
Interal Function for Geocoding with the Batch Endpoint
Please visit the following link for more details:
 \href{https://geocoding.geo.census.gov/geocoder/locations/addressbatch?form}{Census Tract Geocoding}
}
\examples{
\dontrun{
tbd_batch_geocoding(df = data frame)
  }
}
