% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbd_map_mi.R
\name{dynamic_map}
\alias{dynamic_map}
\title{\title{ Dynamic Map Function
}}
\usage{
dynamic_map(
  df,
  map_type,
  pihp_filter,
  cmh_filter,
  col_pallet = "viridis",
  addtiles = "Stamen.TonerLite",
  border_col = "white",
  bins = c(0, 1, 3, 5, 10, 15, 20, 25, 30, 35, 45, 50),
  legend_label = "range"
)
}
\arguments{
\item{df}{A dataframe}

\item{map_type}{A map type ()}
}
\value{

}
\description{
\description{
df : The data frame must contain a name column (county or PIHP or CMHSP names) and summary column i.e... summarized number.
map_type : Basic map with county or PIHP or CMHSP boundaries is drawn based on the map_type argument, Valid values are county, pihp, cmhsp, tract.
pihp_fill : A PIHP filtered dataset must be provided to filter the selected PIHP and highlight the boundaries of the selected PIHP.
cmh_fill : A CMHSP filtered dataset must be provided to filter the selected CMHSP and highlight the boundaries of the selected CMHSP.
col_palette : Choose a color palette of any choice to populate the summarized data.
add tiles : Add a tile layer from a known map provider.
border_col : Fill border color.
bins : Give the bins range in a collection.
legend_label : Label of the legend.
}
}
