% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbd_map_mi.R
\name{static_map}
\alias{static_map}
\title{Static map function}
\usage{
static_map(
  df,
  map_type,
  col_pallet = "viridis",
  addtiles = "Stamen.TonerLite",
  border_col = "white",
  legend_label = "range"
)
}
\arguments{
\item{df}{A dataframe}

\item{ask2}{}
}
\value{

}
\description{
Static map function
}
