% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbd_address_to_tract.R
\name{tbd_address_to_tract}
\alias{tbd_address_to_tract}
\title{Census Tract}
\usage{
tbd_address_to_tract(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
df : A dataframe
}
\examples{
\dontrun{
tbd_address_to_tract(
  df = Data frame
  )
  }
}
