% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbd_address_to_tract.R
\name{tbd_address_to_tract}
\alias{tbd_address_to_tract}
\title{Address to Tract}
\usage{
tbd_address_to_tract(df)
}
\arguments{
\item{df}{A dataframe}
}
\description{
The function allows a dataframe as input and returns geocoded information using
census tract API. The dataframe may not exceed 10,000 records.
}
\examples{
\dontrun{
tbd_address_to_tract(
  df = Data frame
  )
  }
}
